% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.R
\name{myttest}
\alias{myttest}
\title{How to carry out different t-test with a simply function
Takes a list, return the dataframe of two samples with attributes of confident interval and test type}
\usage{
myttest(a, b, alpha, paired)
}
\arguments{
\item{paired}{}
}
\value{
A dataframe consists of two samples and attributes of the confident interval of t -test
}
\description{
How to carry out different t-test with a simply function
Takes a list, return the dataframe of two samples with attributes of confident interval and test type
}
\examples{
set.seed(32)
x=rnorm(30,mean=10,sd=15)
set.seed(35)
y=rnorm(30,mean=8,sd=15)
myttest(x,y,alpha = 0.05, paired = TRUE)
myttest(x,y,alpha = 0.05, paired = FALSE)

}
