% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\title{Get the boxplot for the dataframe using simple function}
\usage{
plot(m)
}
\arguments{
\item{m}{}
}
\value{
boxplot
}
\description{
Get the boxplot for the dataframe using simple function
}
\examples{
plot(x)
}
