% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\title{Get the attributes of the result generated in the t-test.}
\usage{
print(m)
}
\arguments{
\item{m}{}
}
\value{
Attributes of the object
}
\description{
Get the attributes of the result generated in the t-test.
}
\examples{
print(x)
}
